fix: refactor Team component folder structure
